ADD mock redis
