+++ backend/internal/mock_test/mock_cache.go
@@ -1,7 +1,17 @@
 package mock_test
 
-// TODO mockテスト用のredis clientを使用したい
-type MockRedisClient struct{}
+import (
+	"testing"
 
-func (m MockRedisClient) GET(key string, i interface{}) error { return nil }
-func (m MockRedisClient) SET(key string, i interface{}) error { return nil }
+	"github.com/alicebob/miniredis/v2"
+	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/server/cache"
+	"github.com/yoshihiro-shu/draft-backend/backend/internal/config"
+)
+
+func MockRedis(t *testing.T) cache.RedisClient {
+	redisServer := miniredis.RunT(t)
+	return cache.New(config.RedisCache{
+		Host: redisServer.Host(),
+		Port: redisServer.Port(),
+	})
+}
